Move `makeUpdatedStatus` into header `data_type/follow_trajectory.hpp`
