+++ simulation/traffic_simulator/include/traffic_simulator/data_type/follow_trajectory.hpp
@@ -17,8 +17,8 @@
 
 #include <geometry_msgs/msg/pose.hpp>
 #include <optional>
-#include <scenario_simulator_exception/exception.hpp>
-#include <type_traits>
+#include <traffic_simulator_msgs/msg/behavior_parameter.hpp>
+#include <traffic_simulator_msgs/msg/entity_status.hpp>
 
 namespace traffic_simulator
 {
@@ -76,6 +76,13 @@ struct Parameter  // OpenSCENARIO 1.2 FollowTrajectoryAction
   {
   }
 };
+
+auto makeUpdatedStatus(
+  const traffic_simulator_msgs::msg::EntityStatus &,
+  std::shared_ptr<traffic_simulator::follow_trajectory::Parameter<
+    traffic_simulator::follow_trajectory::Polyline>> &,
+  const traffic_simulator_msgs::msg::BehaviorParameter &, double step_time)
+  -> std::optional<traffic_simulator_msgs::msg::EntityStatus>;
 }  // namespace follow_trajectory
 }  // namespace traffic_simulator
